Refactor solve data to be separate of challenges
This removes a hack attaching solve data to the list of challenges,
making them independent. This allows for future expansion, such as
showing a list of solvers for each challenge in the UI.

This commit also removes the `contains` custom template test, as it is
no longer needed by challenges.html.