Disable auto-capitalisation on username input
We disable auto-capitalisation for the username input, so that virtual
keyboards (such as on mobile devices) do not attempt to auto-capitalise
the username, which is likely an annoyance to users.

Note: I discovered that SwiftKey (at least on Android with Chrome)
ignores this attribute, and always auto-capitalises, whereas Gboard
respected the attribute.