Prevent Docker process race condition
Flask's built-in reloader in debug mode can cause multiple processes to
start, each unaware of the other's existence, potentially starting
multiple instances of the same challenge container, which can lead to
port binding failures. To prevent this, we only start Docker processes
when running on first startup in debug mode (always run when not in
debug mode).